+++ src/test/java/io/supertokens/test/bulkimport/BulkImportFlowTest.java
@@ -454,7 +454,6 @@ public class BulkImportFlowTest {
 
         JsonObject response = uploadBulkImportUsersJson(main, usersJson);
         assertEquals("OK", response.get("status").getAsString());
-        System.out.println(response);
 
         // wait for the cron job to process them
         // periodically check the remaining unprocessed users